Move profile to own folder
